// src/app/models/docenteFullInfo.ts
export class Docente {
    identificacion?: number = 0;
    primerNombre?: string = '';
    segundoNombre?: string = '';
    primerApellido?: string = '';
    segundoApellido?: string = '';
    telefono?: string = ''; 
    email?: string = '';
    direccion?: string = '';
    idPrograma?: number = 0;
    enfasis?: string = '';
    estado?: string = '';
    estadoUsuario?: string = '';
    rolDefault?: string = '';
}